+++ tests/test_optimizer_dryruns.py
@@ -2,7 +2,6 @@ import io
 import sys
 import tempfile
 import textwrap
-import time
 from typing import Callable, Optional, Set
 
 from click import testing as cli_testing
@@ -569,31 +568,6 @@ def test_invalid_accelerators_regions(enable_all_clouds):
         sky.stream_and_get(
             sky.launch(task, cluster_name='should-fail', dryrun=True))
         assert 'No launchable resource found for' in str(e.value), str(e.value)
-
-
-def _test_optimize_speed(resources: sky.Resources):
-    with sky.Dag() as dag:
-        task = sky.Task(run='echo hi')
-        task.set_resources(resources)
-    start = time.time()
-    sky.optimize(dag)
-    end = time.time()
-    # 5.0 seconds = somewhat flaky.
-    assert end - start < 6.0, (f'optimize took too long for {resources}, '
-                               f'{end - start} seconds')
-
-
-def test_optimize_speed(enable_all_clouds):
-    _test_optimize_speed(sky.Resources(cpus=4))
-    for cloud in registry.CLOUD_REGISTRY.values():
-        _test_optimize_speed(sky.Resources(infra=str(cloud), cpus='4+'))
-    _test_optimize_speed(sky.Resources(cpus='4+', memory='4+'))
-    _test_optimize_speed(
-        sky.Resources(cpus='4+', memory='4+', accelerators='V100:1'))
-    _test_optimize_speed(
-        sky.Resources(cpus='4+', memory='4+', accelerators='A100-80GB:8'))
-    _test_optimize_speed(
-        sky.Resources(cpus='4+', memory='4+', accelerators='tpu-v3-32'))
 
 
 def test_infer_cloud_from_region_or_zone(enable_all_clouds):
